docs: rename file for typo-name
